+++ sdk/storage/Azure.Storage.DataMovement/tests/Shared/TestTransferWithTimeout.cs
@@ -10,30 +10,14 @@ namespace Azure.Storage.DataMovement.Tests
 {
     public static class TestTransferWithTimeout
     {
-        public static void WaitForCompletion(
-            DataTransfer dataTransfer,
-            TestEventsRaised testEventsRaised,
-            CancellationToken cancellationToken = default)
-        {
-            try
-            {
-                dataTransfer.WaitForCompletion(cancellationToken);
-            }
-            catch (Exception ex)
-            when (ex is OperationCanceledException || ex is TaskCanceledException)
-            {
-                PrintAllEvents(testEventsRaised);
-            }
-        }
-
         public static async Task WaitForCompletionAsync(
-            DataTransfer dataTransfer,
+            TransferOperation transferOperation,
             TestEventsRaised testEventsRaised,
             CancellationToken cancellationToken = default)
         {
             try
             {
-                await dataTransfer.WaitForCompletionAsync(cancellationToken);
+                await transferOperation.WaitForCompletionAsync(cancellationToken);
             }
             catch (Exception ex)
             when (ex is OperationCanceledException || ex is TaskCanceledException)
@@ -83,7 +67,7 @@ namespace Azure.Storage.DataMovement.Tests
                         {
                             Assert.Fail(
                                 $"Status Event at Transfer id: {statusEvent.TransferId}.\n" +
-                                $"Transfer State: {Enum.GetName(typeof(DataTransferState), statusEvent.TransferStatus.State)}\n" +
+                                $"Transfer State: {Enum.GetName(typeof(TransferState), statusEvent.TransferStatus.State)}\n" +
                                 $"HasCompletedSuccessfully: {statusEvent.TransferStatus.HasCompletedSuccessfully}\n" +
                                 $"HasFailedItems: {statusEvent.TransferStatus.HasFailedItems}\n +" +
                                 $"HasSkippedItems: {statusEvent.TransferStatus.HasSkippedItems}\n");